@@ -1,16 +1,16 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 30
+    compileSdkVersion 31
     //buildToolsVersion "28.0.3"
 
     defaultConfig {
-        applicationId "com.hone.black.bravo"
-		minSdkVersion 21
-        targetSdkVersion 29
+        applicationId "com.csv.vpn"
+		minSdkVersion 23
+        targetSdkVersion 30
         versionCode 25
         versionName "3.0"
-		//multiDexEnabled true
+		multiDexEnabled true
     }
     buildTypes {
         release {
@@ -22,9 +22,9 @@ android {
 
 dependencies {
 	// SwetAlert
-	compile 'cn.pedant.sweetalert:library:1.3'
-	
-    compile project(':paygen-ssh-android')
+    implementation 'cn.pedant.sweetalert:library:1.3'
+
+    implementation project(':paygen-ssh-android')
     //api 'com.android.support:multidex:1.0.3'
 	
     //api fileTree(dir: 'libs', include: ['*.jar'])
@@ -35,21 +35,25 @@ dependencies {
 	
 	// support libs
 	
-    compile 'com.android.support:appcompat-v7:25.+'
-	compile 'com.android.support:preference-v7:25.+'
-	compile 'com.android.support:design:25.+'
-	compile 'com.android.support:cardview-v7:+'
+    //noinspection GradleCompatible
+    implementation 'com.android.support:appcompat-v7:25.+'
+    implementation 'com.android.support:preference-v7:25.+'
+    implementation 'com.android.support:design:25.+'
+    implementation 'com.android.support:cardview-v7:+'
+    implementation 'com.google.android.material:material:1.4.0'
+    implementation 'androidx.constraintlayout:constraintlayout:2.1.2'
 	// outros
-	compile 'com.android.support.constraint:constraint-layout:1.0.2'
-	compile 'com.google.android.gms:play-services-ads-lite:11.8.0'
-	compile 'com.github.jetradarmobile:android-snowfall:1.2.0'
+    implementation 'com.android.support.constraint:constraint-layout:1.0.2'
+    implementation 'com.google.android.gms:play-services-ads-lite:11.8.0'
+    implementation 'com.github.jetradarmobile:android-snowfall:1.2.0'
 	//11.8.0
 	//api 'com.scottyab:secure-preferences-lib:0.1.7'
 	//compile 'com.google.android.gms:play-services-ads-lite:15.0.1'
 	//compile 'com.flurry.android:analytics:12.3.0'
 	//api 'com.flurry.android:analytics:12.1.0'
 	//api 'com.android.support:multidex:1.0.3'
-	
-	compile project(':service')
-	compile project(':style')
+
+    implementation project(':service')
+    implementation project(':style')
+    implementation 'androidx.appcompat:appcompat:1.4.0'
 }
