@@ -1,14 +1,12 @@
 apply plugin: 'com.android.library'
 
 android {
-    compileSdkVersion 28
-    buildToolsVersion "28.0.3"
+    compileSdkVersion 29
+    //buildToolsVersion "28.0.3"
 
     defaultConfig {
-        minSdkVersion 21
-        targetSdkVersion 28
-        versionCode 1
-        versionName "1.0"
+        minSdkVersion 23
+        targetSdkVersion 29
     }
     buildTypes {
         release {
@@ -20,18 +18,18 @@ android {
 
 dependencies {
 	//Alert Txt
-	compile 'cn.pedant.sweetalert:library:1.3'
+    api 'cn.pedant.sweetalert:library:1.3'
 	
 	// TastyToast
- compile 'com.sdsmdg.tastytoast:tastytoast:0.1.1'
+    api 'com.sdsmdg.tastytoast:tastytoast:0.1.1'
     
 	//api fileTree(dir: 'libs', include: ['*.jar'])
 	
 	/*api ('android.arch.core:runtime:+') {
         force = true
     }*/
-	
-	compile 'com.android.support:appcompat-v7:25.+'
+
+    api 'com.android.support:appcompat-v7:25.+'
 	
 	//api 'org.connectbot:sshlib:2.2.9'
 	//api 'com.github.SandroMachado:AndroidTampering:1.0.3'
@@ -40,16 +38,17 @@ dependencies {
 	
 	//compile("com.jcraft:jzlib:1.1.3")
     //api 'org.connectbot:simplesocks:1.0.1'
-	
-    compile 'net.i2p.crypto:eddsa:0.3.0'
-	compile 'org.connectbot.jbcrypt:jbcrypt:1.0.0'
+
+    api 'net.i2p.crypto:eddsa:0.3.0'
+    api 'org.connectbot.jbcrypt:jbcrypt:1.0.0'
 	//api 'com.madgag.spongycastle:core:1.58.0.0'
 	
 	// root detect
-	compile 'com.github.kimchangyoun:rootbeerFresh:+'
+    api 'com.github.kimchangyoun:rootbeerFresh:+'
 	
 	//api 'com.github.hussainderry:secure-preferences:4.0.0'
-	
-	compile project(':style')
-	
+
+    api project(':style')
+    implementation 'com.google.android.material:material:1.4.0'
+
 }
