@@ -1,7 +1,7 @@
 package com.csv.vpn.model;
 
 import com.csv.vpn.SocksHttpMainActivity;
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 import android.view.View;
 
 public abstract class ViewFragment extends Fragment
