@@ -4,7 +4,7 @@ import android.content.Intent;
 import android.os.Build;
 import android.content.Context;
 import com.slipkprojects.ultrasshservice.SocksHttpService;
-import android.support.v4.content.LocalBroadcastManager;
+import androidx.localbroadcastmanager.content.LocalBroadcastManager;
 import com.slipkprojects.ultrasshservice.config.Settings;
 
 public class TunnelManagerHelper
