@@ -3,12 +3,13 @@ package com.slipkprojects.ultrasshservice.tunnel;
 import android.content.Context;
 import android.net.ConnectivityManager;
 import android.net.NetworkInfo;
+import androidx.collection.ArrayMap;
 import com.slipkprojects.ultrasshservice.logger.SkStatus;
 import java.io.UnsupportedEncodingException;
 import java.util.regex.Pattern;
 import java.io.OutputStream;
 import java.io.IOException;
-import android.support.v4.util.ArrayMap;
+import androidx.collection.ArrayMap;
 import java.util.Map;
 import java.util.Enumeration;
 import java.net.NetworkInterface;
