@@ -6,7 +6,7 @@
 package com.slipkprojects.ultrasshservice.tunnel.vpn;
 
 import android.os.Build;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import java.math.BigInteger;
 import java.net.Inet6Address;
@@ -15,6 +15,7 @@ import java.util.Locale;
 import java.util.PriorityQueue;
 import java.util.TreeSet;
 import java.util.Vector;
+
 import com.slipkprojects.ultrasshservice.BuildConfig;
 
 
