@@ -7,11 +7,13 @@ import android.util.Log;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import android.content.Context;
+import androidx.annotation.NonNull;
+import androidx.annotation.RequiresApi;
+import androidx.localbroadcastmanager.content.LocalBroadcastManager;
 import com.slipkprojects.ultrasshservice.logger.SkStatus;
 import android.os.Binder;
 import android.os.Handler;
 import android.content.IntentFilter;
-import android.support.v4.content.LocalBroadcastManager;
 import java.util.List;
 import com.slipkprojects.ultrasshservice.aidl.IUltraSSHServiceInternal;
 import android.os.RemoteException;
@@ -19,14 +21,12 @@ import android.annotation.TargetApi;
 import android.os.Build;
 import android.app.Notification;
 import com.slipkprojects.ultrasshservice.logger.ConnectionStatus;
-import android.support.annotation.NonNull;
 import android.app.NotificationManager;
 import java.lang.reflect.Method;
 import java.lang.reflect.InvocationTargetException;
 import android.app.PendingIntent;
 import android.content.ComponentName;
 import java.io.IOException;
-import android.support.annotation.RequiresApi;
 import com.slipkprojects.ultrasshservice.tunnel.TunnelUtils;
 import com.slipkprojects.ultrasshservice.tunnel.TunnelManagerThread;
 import android.content.BroadcastReceiver;
